app/requirements.txt
Flask==1.1.0
requests==2.22.0
mysql-connector-python-rf
apiclient
oauth2client
google-api-python-client
# sqlalchemy
# psycopg2
#gunicorn==19.9.0
#Keras>=2.2.5
#numpy==1.16.4
#pandas==0.24.2
#pillow>=6.2.0
#python-dateutil==2.8.0
#pytz==2019.1
#PyYAML==5.1.1
#scikit-learn==0.21.2
#sklearn==0.0
#matplotlib==3.1.1
#tensorboard>=1.14.0
# tensorflow>=1.14.0
#mecab-python3==0.996.2